Cleaned out a bunch of crap in the .gitignore.
Made it make more sense.  Linux/Unix temp files, generated content (i.e.
the game binary, log files, and the build directory.  Removed the stupid
(and yes, I'm calling it that) crap for .Net which shouldn't have been
a "thing" with the previous "owner" of the project because he *NEVER*
finished the Pascal fixes and made the game work, blaming "real life"
as the reason.  And I've gotten a bit of the bitrot back out and got
it started.