Removed BinkPlayer.
Don't need it.  Don't have a license for redistribution and while
RAD provides it for "free", we're not keeping it around.  VLC
does at least as well and there's other options.  We'll be
wiring VLC into the path for now and make it an option for install
for the FOSS version of this game.