Clobbered the "old" source.
Kept the video playback piece.  VLC is "fine" for now.  The jedi-sdl
version may have a few 'good' items, but unlikely.  Port process from
my predecessors in the "Team SOAOS" group started from wrong direction
in light of too many key pieces (Like the path finder using Win16(!!)
code calls...Win16 calls in an engine that was released in 2000?) not
being properly handled or even STARTED.  So...we're starting from
"ground zero" and re-wiring SDL2 into the beastie and making DirectX
like effects properly instead of simply writing them off like the
Team SOAOS "contributor" opted for.

No, I'm not Team SOAOS.  It's LGPLed.  SO...I'm going to, like the
README.md says, I'm making a fresh go at this with an eye on keeping
the Delphi-ish sources, using FPC/Lazarus which have what it takes
to actually make this happen on multi[ple targets.

With that in mind, I grabbed the Win32 directory, dropped it in and
started the process for moving it to Lazarus, SDL, etc.  Not complete
right now, but so far so good- it looks like I've got the AI path
management stuff moved over better than not at all and close to
a proper form for a modern version of all of this that is actually
Cross-Platform.