Cleaned up .gitignore further.
Took the reference to src/jedi-sdl/dcu out of the ignores (Won't be
there ever again.  When I do automation for multiple Arch and OS
targets, we'll revisit this with an eye on doing it "better" than
the prior maintainers did.)"

Added Lazarus-centric ignores for the target.  Once we get a bit
further along, we might need to amend/change the executables
entries in that- but for now...