Phase one in an initial attempt to modernize and SDL2-ize code
Removed "DirectX" from all of the uses clauses in the project.  Don't
need them once we move to SDL2.

Removed mention of "Windows" from the uses clasues in the project.  We
are going to try to not need 'Windows' as a requirement for anything in
this re-rendition of the game.  SDL or something like it for whatever
else we need that is truly cross-platform will do for that.

Added the START of SDL2-ization.  Not even close to done, just did some
sed substitutions on one of the files.  Need to actually do the
implementations according to how one would do it in SDL.  Result is
that we slowly replace DX/D3D calls with the corresponding SDL/OGL
code underneath and as we come fully up, we refactor the naming/code
to be cross-platform.

I've got the old Windows code in the other directory for now, so
I've got a crib-sheet to nab and re-attempt some of this if I
bust something TOO badly- but I need to drop this non-compiling
stuff as a WIP into Git so I have a safe place globally to fetch
the WIP from.