Added SDL2 into the mix for stage one of the re-work
For now, doing SDL2 and a move TO it.  The "work" done to this
point on SOAOS has been...iffy...   Not to put too fine of a point
on any of this, the code got started and while there's claims of
having it work on Linux as claimed is...dubious at best.  There
is a **LOT** of DirectX-isms and DirectDraw-isms that should've
not been there in this codebase if it was as claimed.

I'll forge forward for now.  It may be that the engine algorithms
and the assets are the only salvageable things from the "Siege Engine"
at this time because everything's moved on and this, while the
game itself is probably worth saving, the guts of it with all the
bitrot and cruft from Win16 and Win32 coding that should probably
have not been in a title from 2000 of all things... (Seriously?)