General cleanups.
Moved much of it to refer to TSDL_surface instead of Direct Draw for
variables using DirectDraw surfaces.  Not QUITE the right thing to do
overall, I suspect, but it's a start in the right direction.  It
still doesn't fully compile, but it's a good start in the relative
right direction.  I can work on blits, moves of the DirectDraw crap
out of the code and SDL2 based calls into it.